requirements.txt
ccxt
pandas
pandas_ta
scikit-learn
joblib
python-telegram-bot
